Moved logout view tests to mixin
This mixin can be used by the IDAs to test their implementations of the logout view.

ECOM-4607
