+++ auth_backends/__init__.py
@@ -3,4 +3,4 @@
  These package is designed to be used primarily with Open edX Django projects, but should be compatible with non-edX
  projects as well.
 """
-__version__ = '0.3.0'  # pragma: no cover
+__version__ = '0.3.1'  # pragma: no cover
